requirements.txt
pyautogui
pydirectinput
keyboard
opencv-python
numpy
pytesseract
pillow
